agents/adk_nextjs/pyproject.toml
[project]
name = "adk-a2a-agent-adk-nextjs"
version = "0.1.0"
description = "Google ADK-based Expense Reimbursement agent hosted as an A2A server." 
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adk-a2a-lib",
    "click>=8.1.8",
    "google-adk>=0.0.3",
    "google-genai>=1.9.0",
    "python-dotenv>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
adk-a2a-lib = { path = "../../lib", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
